Revert "本地更新cpp文件"
This reverts commit 74a9d3cdf1d044d0e04afbe5da71f80bcb0de1f0.

+++ app/src/main/java/com/luck/pictureselector/newtestfile.cpp
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main()
 {
-    int d,e,f;
+    float d,e,f;
     int a,b,c;
     printf("hello world!\n");
     return 0;
